move package endpoints to separate module
