+++ ledger-service/http-json/src/main/scala/com/digitalasset/http/endpoints/RouteSetup.scala
@@ -19,10 +19,11 @@ import com.codahale.metrics.Timer
 import com.daml.logging.LoggingContextOf.withEnrichedLoggingContext
 import com.daml.metrics.Metrics
 import com.daml.scalautil.Statement.discard
-import domain.{JwtPayloadG, JwtPayloadTag, JwtWritePayload}
+import domain.{JwtPayloadG, JwtPayloadLedgerIdOnly, JwtPayloadTag, JwtWritePayload}
 import json._
 import util.FutureUtil.{either, eitherT}
 import util.Logging.{InstanceUUID, RequestID}
+import util.toLedgerId
 import com.daml.jwt.domain.Jwt
 import com.daml.ledger.api.{v1 => lav1}
 import lav1.value.{Value => ApiValue}
@@ -35,6 +36,7 @@ import scala.concurrent.duration.FiniteDuration
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.Try
 import scala.util.control.NonFatal
+import com.daml.ledger.api.{domain => LedgerApiDomain}
 import com.daml.ledger.client.services.admin.UserManagementClient
 import com.daml.ledger.client.services.identity.LedgerIdentityClient
 import com.daml.logging.{ContextualizedLogger, LoggingContextOf}
@@ -68,6 +70,16 @@ private[http] final class RouteSetup(
   )(req: HttpRequest)(implicit
       lc: LoggingContextOf[InstanceUUID with RequestID]
   ): ET[R] = proxyWithCommand((jwt, a: A) => fn(jwt, a).run)(req)
+
+  def proxyWithoutCommand[A](
+      fn: (Jwt, LedgerApiDomain.LedgerId) => Future[A]
+  )(req: HttpRequest)(implicit
+      lc: LoggingContextOf[InstanceUUID with RequestID]
+  ): ET[A] =
+    for {
+      t3 <- inputAndJwtPayload[JwtPayloadLedgerIdOnly](req).leftMap(it => it: Error)
+      a <- eitherT(handleFutureFailure(fn(t3._1, toLedgerId(t3._2.ledgerId)))): ET[A]
+    } yield a
 
   private[endpoints] def handleCommand[T[_]](req: HttpRequest)(
       fn: (
@@ -110,6 +122,15 @@ private[http] final class RouteSetup(
     decodeAndParsePayload[P](fa._1, decodeJwt, userManagementClient, ledgerIdentityClient).map(t2 =>
       (t2._1, t2._2, fa._2)
     )
+
+  def inputAndJwtPayload[P](
+      req: HttpRequest
+  )(implicit
+      legacyParse: ParsePayload[P],
+      createFromUserToken: CreateFromUserToken[P],
+      lc: LoggingContextOf[InstanceUUID with RequestID],
+  ): EitherT[Future, Unauthorized, (Jwt, P, String)] =
+    eitherT(input(req)).flatMap(it => withJwtPayload[String, P](it))
 
   def getParseAndDecodeTimerCtx()(implicit
       metrics: Metrics
@@ -187,6 +208,12 @@ private[http] object RouteSetup {
       "application_id" -> jwtPayload.applicationId.toString,
       "read_as" -> jwtPayload.readAs.toString,
     ).run(fn)
+
+  private[http] def handleFutureFailure[A](fa: Future[A])(implicit
+      ec: ExecutionContext,
+      lc: LoggingContextOf[InstanceUUID with RequestID],
+  ): Future[Error \/ A] =
+    fa.map(a => \/-(a)).recover(logException("Future") andThen Error.fromThrowable andThen (-\/(_)))
 
   private[endpoints] def handleFutureEitherFailure[A, B](fa: Future[A \/ B])(implicit
       ec: ExecutionContext,
